Logging improvements: - Moved logging from an object to a static library; - Removed deprecated auto_ptr, as we moved to a latest version of log4cplus; - All submodules are implicitly linked against logging library; - Split iternal and log4cplus implementations into separate files; - Force default log4cplus library initialization.
