+++ bftengine/src/preprocessor/tests/CMakeLists.txt
@@ -1,6 +1,6 @@
 find_package(GTest REQUIRED)
 
-add_executable(preprocessor_test preprocessor_test.cpp $<TARGET_OBJECTS:logging_dev>)
+add_executable(preprocessor_test preprocessor_test.cpp )
 add_test(preprocessor_test preprocessor_test)
 
 target_include_directories(preprocessor_test PUBLIC ..)
