+++ bftengine/tests/controllerWithSimpleHistory/CMakeLists.txt
@@ -1,6 +1,6 @@
 find_package(GTest REQUIRED)
 
-add_executable(ControllerWithSimpleHistory_test ControllerWithSimpleHistory_test.cpp $<TARGET_OBJECTS:logging_dev>)
+add_executable(ControllerWithSimpleHistory_test ControllerWithSimpleHistory_test.cpp )
 add_test(ControllerWithSimpleHistory_test ControllerWithSimpleHistory_test)
 
 target_link_libraries(ControllerWithSimpleHistory_test PUBLIC
