+++ bftengine/tests/metadataStorage/CMakeLists.txt
@@ -1,6 +1,6 @@
 if (BUILD_ROCKSDB_STORAGE)
     find_package(GTest REQUIRED)
-    add_executable(metadataStorage_test metadataStorage_test.cpp $<TARGET_OBJECTS:logging_dev>)
+    add_executable(metadataStorage_test metadataStorage_test.cpp )
     add_test(metadataStorage_test metadataStorage_test)
 
     target_link_libraries(metadataStorage_test PUBLIC
