+++ bftengine/tests/simpleStorage/CMakeLists.txt
@@ -11,6 +11,6 @@ add_executable(simple_file_storage ${simple_file_storage_source_files})
 
 add_library(simple_file_storage_lib ${simple_file_storage_source_files_lib})
 
-target_link_libraries(simple_file_storage PUBLIC corebft  $<TARGET_OBJECTS:logging_dev>)
+target_link_libraries(simple_file_storage PUBLIC corebft  )
 
 target_link_libraries(simple_file_storage_lib PUBLIC corebft)
