+++ bftengine/tests/testSeqNumForClientRequest/CMakeLists.txt
@@ -1,6 +1,6 @@
 find_package(GTest REQUIRED)
 
-add_executable(seqNumForClientRequest_test seqNumForClientRequest_test.cpp $<TARGET_OBJECTS:logging_dev>)
+add_executable(seqNumForClientRequest_test seqNumForClientRequest_test.cpp )
 add_test(seqNumForClientRequest_test seqNumForClientRequest_test)
 
 target_include_directories(seqNumForClientRequest_test
