+++ bftengine/tests/testSerialization/CMakeLists.txt
@@ -9,5 +9,5 @@ add_executable(test_serialization
 target_include_directories(test_serialization PUBLIC
     ${bftengine_SOURCE_DIR}/src/bftengine)
 
-target_link_libraries(test_serialization PUBLIC corebft $<TARGET_OBJECTS:logging_dev>)
+target_link_libraries(test_serialization PUBLIC corebft )
 add_test(NAME test_serialization COMMAND test_serialization)
