+++ kvbc/benchmark/CMakeLists.txt
@@ -5,7 +5,7 @@
 find_package(benchmark QUIET)
 
 if(benchmark_FOUND)
-    add_executable(sparse_merkle_benchmark sparse_merkle_benchmark.cpp $<TARGET_OBJECTS:logging_dev>)
+    add_executable(sparse_merkle_benchmark sparse_merkle_benchmark.cpp )
     target_link_libraries(sparse_merkle_benchmark PUBLIC
         benchmark
         util
