+++ kvbc/test/CMakeLists.txt
@@ -2,7 +2,7 @@ find_package(GTest REQUIRED)
 find_package(rapidcheck)
 
 if(Boost_FOUND)
-    add_executable(blockchain_view_test blockchain_view_test.cpp $<TARGET_OBJECTS:logging_dev>)
+    add_executable(blockchain_view_test blockchain_view_test.cpp )
     add_test(blockchain_view_test blockchain_view_test)
     target_include_directories(blockchain_view_test PRIVATE ${Boost_INCLUDE_DIRS})
     target_link_libraries(blockchain_view_test PUBLIC
@@ -12,18 +12,18 @@ if(Boost_FOUND)
     )
 endif()
 
-add_executable(order_test order_test.cpp $<TARGET_OBJECTS:logging_dev>)
+add_executable(order_test order_test.cpp )
 add_test(order_test order_test)
 target_link_libraries(order_test GTest::Main kvbc util)
 target_compile_options(order_test PUBLIC -Wno-sign-compare)
 
-add_executable(kvbc_dbadapter_test kvbc_dbadapter_test.cpp $<TARGET_OBJECTS:logging_dev>)
+add_executable(kvbc_dbadapter_test kvbc_dbadapter_test.cpp )
 add_test(kvbc_dbadapter_test kvbc_dbadapter_test)
 target_link_libraries(kvbc_dbadapter_test GTest::Main kvbc util)
 
 
 add_executable(sparse_merkle_storage_db_adapter_unit_test
-    sparse_merkle_storage/db_adapter_unit_test.cpp $<TARGET_OBJECTS:logging_dev>)
+    sparse_merkle_storage/db_adapter_unit_test.cpp )
 add_test(sparse_merkle_storage_db_adapter_unit_test sparse_merkle_storage_db_adapter_unit_test)
 target_link_libraries(sparse_merkle_storage_db_adapter_unit_test PUBLIC
     GTest::Main
@@ -35,7 +35,7 @@ target_link_libraries(sparse_merkle_storage_db_adapter_unit_test PUBLIC
 )
 
 add_executable(sparse_merkle_storage_serialization_unit_test
-    sparse_merkle_storage/serialization_unit_test.cpp $<TARGET_OBJECTS:logging_dev>)
+    sparse_merkle_storage/serialization_unit_test.cpp )
 add_test(sparse_merkle_storage_serialization_unit_test sparse_merkle_storage_serialization_unit_test)
 target_link_libraries(sparse_merkle_storage_serialization_unit_test PUBLIC
     GTest::Main
@@ -46,7 +46,7 @@ target_link_libraries(sparse_merkle_storage_serialization_unit_test PUBLIC
 
 if(RAPIDCHECK_FOUND)
 add_executable(sparse_merkle_storage_db_adapter_property_test
-    sparse_merkle_storage/db_adapter_property_test.cpp $<TARGET_OBJECTS:logging_dev>)
+    sparse_merkle_storage/db_adapter_property_test.cpp )
 add_test(sparse_merkle_storage_db_adapter_property_test sparse_merkle_storage_db_adapter_property_test)
 # Decrease both the input size and test count from the default value of 100 to 50 in order to speed up this test.
 set_property(TEST sparse_merkle_storage_db_adapter_property_test PROPERTY ENVIRONMENT RC_PARAMS=max_size=50\ max_success=50)
@@ -62,7 +62,7 @@ target_link_libraries(sparse_merkle_storage_db_adapter_property_test PUBLIC
     )
 
 add_executable(sparse_merkle_internal_node_property_test sparse_merkle/internal_node_property_tests.cpp
-$<TARGET_OBJECTS:logging_dev>)
+)
 add_test(sparse_merkle_internal_node_property_test sparse_merkle_internal_node_property_test)
 target_include_directories(sparse_merkle_internal_node_property_test PRIVATE ${RAPIDCHECK_INCLUDE_DIRS})
 target_link_libraries(sparse_merkle_internal_node_property_test PUBLIC
@@ -76,7 +76,7 @@ target_link_libraries(sparse_merkle_internal_node_property_test PUBLIC
 endif ()
 
 add_executable(sparse_merkle_base_types_test sparse_merkle/base_types_test.cpp
-    $<TARGET_OBJECTS:logging_dev>)
+    )
 add_test(sparse_merkle_base_types_test sparse_merkle_base_types_test)
 target_link_libraries(sparse_merkle_base_types_test PUBLIC
     GTest::Main
@@ -87,7 +87,7 @@ target_link_libraries(sparse_merkle_base_types_test PUBLIC
 
 find_package(OpenSSL REQUIRED)
 add_executable(sparse_merkle_internal_node_test sparse_merkle/internal_node_test.cpp
-    $<TARGET_OBJECTS:logging_dev>)
+    )
 add_test(sparse_merkle_internal_node_test sparse_merkle_internal_node_test)
 target_link_libraries(sparse_merkle_internal_node_test PUBLIC
     GTest::Main
@@ -99,7 +99,7 @@ target_link_libraries(sparse_merkle_internal_node_test PUBLIC
 
 find_package(OpenSSL REQUIRED)
 add_executable(sparse_merkle_tree_test sparse_merkle/tree_test.cpp
-    $<TARGET_OBJECTS:logging_dev>)
+    )
 add_test(sparse_merkle_tree_test sparse_merkle_tree_test)
 target_link_libraries(sparse_merkle_tree_test PUBLIC
     GTest::Main
@@ -110,7 +110,7 @@ target_link_libraries(sparse_merkle_tree_test PUBLIC
 )
 
 if (BUILD_ROCKSDB_STORAGE)
-add_executable(multiIO_test multiIO_test.cpp $<TARGET_OBJECTS:logging_dev>)
+add_executable(multiIO_test multiIO_test.cpp )
 add_test(multiIO_test multiIO_test)
 
 target_link_libraries(multiIO_test PUBLIC
