+++ tests/simpleKVBC/TesterReplica/CMakeLists.txt
@@ -19,7 +19,7 @@ if(BUILD_ROCKSDB_STORAGE)
 	target_compile_definitions(skvbc_replica PUBLIC "USE_ROCKSDB=1")
 endif()
 
-target_link_libraries(skvbc_replica PUBLIC kvbc corebft threshsign util test_config_lib $<TARGET_OBJECTS:logging_dev>)
+target_link_libraries(skvbc_replica PUBLIC kvbc corebft threshsign util test_config_lib )
 
 target_include_directories(skvbc_replica PUBLIC ..)
 target_include_directories(skvbc_replica PUBLIC ../..)
