+++ util/test/CMakeLists.txt
@@ -6,7 +6,7 @@ add_test(metric_tests metric_tests)
 target_link_libraries(metric_tests GTest::Main util)
 target_compile_options(metric_tests PUBLIC -Wno-sign-compare)
 
-add_executable(metric_server MetricServerTestMain.cpp $<TARGET_OBJECTS:logging_dev>)
+add_executable(metric_server MetricServerTestMain.cpp )
 target_link_libraries(metric_server util)
 
 add_test(NAME metric_server_tests COMMAND python3 -m unittest
@@ -15,28 +15,28 @@ add_test(NAME metric_server_tests COMMAND python3 -m unittest
 
 add_executable(mt_tests multithreading.cpp)
 add_test(util_mt_tests mt_tests)
-target_link_libraries(mt_tests GTest::Main util $<TARGET_OBJECTS:logging_dev>)
+target_link_libraries(mt_tests GTest::Main util )
 target_compile_options(mt_tests PUBLIC -Wno-sign-compare)
 
-add_executable(sliver_test sliver_test.cpp $<TARGET_OBJECTS:logging_dev>)
+add_executable(sliver_test sliver_test.cpp )
 add_test(sliver_test sliver_test)
 target_link_libraries(sliver_test GTest::Main util)
 target_compile_options(sliver_test PUBLIC -Wno-sign-compare)
 
-add_executable(serializable_test serializable_test.cpp $<TARGET_OBJECTS:logging_dev>)
+add_executable(serializable_test serializable_test.cpp )
 add_test(serializable_test serializable_test)
 target_link_libraries(serializable_test GTest::Main util)
 target_compile_options(serializable_test PUBLIC -Wno-sign-compare)
 
-add_executable(timers_tests timers_tests.cpp $<TARGET_OBJECTS:logging_dev>)
+add_executable(timers_tests timers_tests.cpp )
 add_test(timers_tests timers_tests)
 target_link_libraries(timers_tests GTest::Main util)
 
 add_executable(sha3_256_tests sha3_256_tests.cpp)
 add_test(sha3_256_tests sha3_256_tests)
 target_link_libraries(sha3_256_tests GTest::Main util OpenSSL::Crypto
-        $<TARGET_OBJECTS:logging_dev>)
+        )
 
-add_executable(RollingAvgAndVar_test RollingAvgAndVar_test.cpp $<TARGET_OBJECTS:logging_dev>)
+add_executable(RollingAvgAndVar_test RollingAvgAndVar_test.cpp )
 add_test(RollingAvgAndVar_test RollingAvgAndVar_test)
 target_link_libraries(RollingAvgAndVar_test GTest::Main util)
